+++ mopidy/local/json.py
@@ -61,8 +61,7 @@ class _BrowseCache(object):
     splitpath_re = re.compile(r'([^/]+)')
 
     def __init__(self, uris):
-        # TODO: local.ROOT_DIRECTORY_URI
-        self._cache = {'local:directory': collections.OrderedDict()}
+        self._cache = {local.ROOT_DIRECTORY_URI: collections.OrderedDict()}
 
         for track_uri in uris:
             path = translator.local_track_uri_to_path(track_uri, b'/')
@@ -97,10 +96,10 @@ class _BrowseCache(object):
             else:
                 # Loop completed, so final child needs to be added to root.
                 if child:
-                    self._cache['local:directory'][child.uri] = child
+                    self._cache[local.ROOT_DIRECTORY_URI][child.uri] = child
                 # If no parent was set we belong in the root.
                 if not parent_uri:
-                    parent_uri = 'local:directory'
+                    parent_uri = local.ROOT_DIRECTORY_URI
 
             self._cache[parent_uri][track_uri] = track_ref